fix(middleware): fix bug in clearChat middleware

+++ TaskyBot/src/middlewares/clearChat.ts
@@ -2,13 +2,16 @@ import { NextFunction } from 'grammy';
 import { MyContext } from '../bot.js';
 
 export async function clearChat(ctx: MyContext, next: NextFunction) {
-  if (ctx.message && ctx.session.__last_bot_message) {
+  const userAction = ctx.message || ctx.callbackQuery;
+
+  if (userAction && ctx.session.__last_bot_message) {
     try {
-      if (ctx.chat?.id)
+      if (!ctx.session.__last_command_start && ctx.chat?.id) {
         await ctx.api.deleteMessage(
-          ctx.chat?.id,
+          ctx.chat.id,
           ctx.session.__last_bot_message,
         );
+      }
     } catch (err) {
       console.error('❌ Error deleting message:', err);
     }
@@ -23,4 +26,10 @@ export async function clearChat(ctx: MyContext, next: NextFunction) {
   };
 
   await next();
+
+  if (ctx.message?.text === '/start') {
+    ctx.session.__last_command_start = true;
+  } else {
+    ctx.session.__last_command_start = false;
+  }
 }
